Bugfix for Connectivity Issues
This commit ensures that the NLB and ASG are in the same subnet (so that traffic from
the listeners is not directed at an interface that doesn't exist). It also ensures that
the EC2 instance for the ASG has a public IP so that it can hit external targets (this
is necessary for cloud-based LDAP solutions).